tests/test_auth.py
import unittest
from app import app
from auth import init_db
from models import db  # Assuming you have a db object in your models module

class TestAuth(unittest.TestCase):
    def setUp(self):
        app.config['TESTING'] = True
        self.app = app.test_client()

    def tearDown(self):
        # Clean up database and application context after each test
        with app.app_context():
            db.drop_all()

    def test_database_initialization(self):
        with app.app_context():
            init_db(app)
            # Perform assertions to check if the database has been initialized as expected
            self.assertTrue(db.engine.table_names())
            # Add more assertions as needed

if __name__ == '__main__':
    unittest.main()